+++ src/textual/widgets/_label.py
@@ -5,3 +5,11 @@ from ._static import Static
 
 class Label(Static):
     """A simple label widget for displaying text-oriented renderables."""
+
+    DEFAULT_CSS = """
+    Static {
+        width: auto;
+        height: auto;
+    }
+    """
+    """str: The default styling of a `Label`."""